+++ changes.md
@@ -1,5 +1,10 @@
 # JSROOT changelog
 
+## Changes in master
+1. Try to provide workaround for websites where requirejs already loaded.
+   This makes problem by direct loading of jquery and jquery-ui 
+
+
 ## Changes in 3.5
 1. Fix error in vertical text alignment
 2. Many improvements in TPaletteAxis drawing - draw label, avoid too large ticks.